Move test_automl iterative tests to test_automl_iterative_algorithm.py
